+++ tests/unit/tasks/Generate_Wiki_Content.build.Tests.ps1
@@ -33,7 +33,7 @@ Describe 'Generate_Wiki_Content' {
         }
     }
 
-    It 'Should run the build script alias' {
+    It 'Should export the build script alias' {
         $buildTaskName = 'Generate_Wiki_Content'
         $buildScriptAliasName = 'Task.{0}' -f $buildTaskName
 
@@ -43,8 +43,8 @@ Describe 'Generate_Wiki_Content' {
         $script:buildScript.ReferencedCommand | Should -Be ('{0}.build.ps1' -f $buildTaskName)
     }
 
-    It 'Should dot-source the build script without throwing' {
-        { . $script:buildScript } | Should -Not -Throw
+    It 'Should reference an existing build script' {
+        Test-Path -Path $script:buildScript.Definition | Should -BeTrue
     }
 
     It 'Should run the build task without throwing' {